looping works! need to make the flag values more precise than 1sec. should eventually do error validation with 0.1 precision for example

+++ MusicPlayerForDrummers/View/Controls/Player/WaveformSeekbar.xaml.cs
@@ -228,9 +228,9 @@ namespace MusicPlayerForDrummers.View.Controls.Player
                 else if (_flag == EndScrollFlag)
                     playerVM.Session.PlayingSong.ScrollEndTime = (int) Math.Floor(playerVM.Session.Player.Length * Canvas.GetRight(_flag) / FlagsCanvas.ActualWidth);
                 else if (_flag == StartLoopFlag)
-                    playerVM.Session.Player.LoopStart = (double) Math.Floor(playerVM.Session.Player.Length * Canvas.GetLeft(_flag) / FlagsCanvas.ActualWidth);
+                    playerVM.Session.Player.LoopStart = Math.Floor(playerVM.Session.Player.Length * Canvas.GetLeft(_flag) / FlagsCanvas.ActualWidth);
                 else if (_flag == EndLoopFlag)
-                    playerVM.Session.Player.LoopEnd = (double) Math.Floor(playerVM.Session.Player.Length * Canvas.GetLeft(_flag) / FlagsCanvas.ActualWidth);
+                    playerVM.Session.Player.LoopEnd = Math.Floor(playerVM.Session.Player.Length * Canvas.GetLeft(_flag) / FlagsCanvas.ActualWidth);
 
                 _flag = null;
             }
